requirements.txt
anyascii==0.3.1
asgiref==3.5.0
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-filter==21.1
django-modelcluster==5.3
django-recaptcha==3.0.0
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
openpyxl==3.0.9
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
urllib3==1.26.9
wagtail==2.16.2
wagtail-django-recaptcha==1.0
wagtailmedia==0.9.0
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
